ca2reforco/requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.2
django-admin-rangefilter==0.9.0
django-filter==22.1
Pillow==9.3.0
pycodestyle==2.9.1
simplejson==3.18.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.5
